mutualfunds-stock-mapping/requirements.txt
scrapy==2.11.1
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
validators==0.20.0
streamlit==1.32.0
nest-asyncio==1.6.0
lxml==5.1.0
aiohttp==3.9.3 